Refactor userRepo and add utils file to hold helper methods
